feat: drab, dull ping responses a thing of the past (#27)
Instead we get shiny new pet names for our pet services! These are randomly generated at process startup using the `petname` crate. The string is also included on a new `ping` field in the response struct for `/monitor/status`, in case you might find it handy to correlate.

Updated all the tests that check for the ping response to verify that the response is non-empty instead of checking for the static service name.